requirements.txt
absl-py==0.8.1
astor==0.8.0
bcrypt==3.1.7
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
numpy==1.17.3
opencv-python==4.1.1.26
opt-einsum==3.1.0
protobuf==3.10.0
pycparser==2.19
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
